refactor: ADR-004 Phase 5-E & 5-F - like/donation 패키지 application 계층 이관
## Phase 5-E: Like Package Migration
- CharacterLikeService → application/service/like/
- LikeSyncEventListener → application/service/like/listener/
- LikeSyncMetricsRecorder → application/service/like/metrics/
- OrphanKeyRecoveryService → application/service/like/recovery/
- 삭제: LikeSyncFailedEvent.java (Kotlin 버전 사용: module-infra)
- 삭제: AtomicFetchStrategy.java (Kotlin 인터페이스 사용: LikeAtomicFetchStrategy.kt)

## Phase 5-F: Donation Package Migration
- PaymentStrategy → application/service/donation/
- InternalPointPaymentStrategy → application/service/donation/
- DonationProcessor → application/service/donation/
- DonationFailedEvent → application/service/donation/event/
- DonationEventListener → application/service/donation/listener/
- DlqAdminService → application/service/donation/outbox/

## Import Updates
- LikeSyncService: 새 위치로 import 업데이트
- DonationService: 새 위치로 import 업데이트
- 테스트 파일 import 업데이트
