+++ module-app/src/main/java/maple/expectation/application/service/donation/DonationProcessor.java
@@ -1,8 +1,7 @@
-package maple.expectation.service.v2.donation.event;
+package maple.expectation.application.service.donation;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import maple.expectation.service.v2.donation.PaymentStrategy;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
