+++ module-app/src/main/java/maple/expectation/service/v2/LikeSyncService.java
@@ -9,6 +9,7 @@ import java.util.UUID;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.concurrent.atomic.AtomicLong;
 import lombok.extern.slf4j.Slf4j;
+import maple.expectation.application.service.like.metrics.LikeSyncMetricsRecorder;
 import maple.expectation.core.dto.like.FetchResult;
 import maple.expectation.core.port.out.LikeBufferStrategy;
 import maple.expectation.core.port.out.LikeSyncPort;
@@ -21,7 +22,6 @@ import maple.expectation.infrastructure.executor.TaskContext;
 import maple.expectation.infrastructure.queue.like.LikeSyncExecutor;
 import maple.expectation.infrastructure.queue.like.compensation.RedisCompensationCommand;
 import maple.expectation.infrastructure.shutdown.dto.FlushResult;
-import maple.expectation.service.v2.like.metrics.LikeSyncMetricsRecorder;
 import maple.expectation.service.v2.shutdown.ShutdownDataPersistenceService;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.ApplicationEventPublisher;
