+++ module-app/src/test/java/maple/expectation/service/v2/donation/outbox/DlqAdminServiceTest.java
@@ -10,6 +10,7 @@ import static org.mockito.Mockito.*;
 
 import java.util.List;
 import java.util.Optional;
+import maple.expectation.application.service.donation.outbox.DlqAdminService;
 import maple.expectation.domain.v2.DonationDlq;
 import maple.expectation.domain.v2.DonationOutbox;
 import maple.expectation.error.exception.DlqNotFoundException;
